pyproject.toml
[tool.poetry]
name = "gh-llm-recup"
version = "0.1.0"
description = ""
authors = ["kolommik <mikhail.kolomasov@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.40.2"
openai = "^1.55.3"
python-dotenv = "^1.0.1"
tiktoken = "^0.8.0"
anthropic = "^0.40.0"
toml = "^0.10.2"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
mypy = "^1.13.0"
pylint = "^3.3.1"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
